NATIVE-42: Move events mock to own file
